tests pass and update config.yml for circleci
