+++ contracts/sg-eth-airdrop/src/lib.rs
@@ -6,11 +6,10 @@ pub mod reply;
 pub mod state;
 pub use crate::error::ContractError;
 
-#[path = "./helpers/lib.rs"]
-pub mod helpers;
-
 #[path = "./ethereum/lib.rs"]
 mod ethereum;
+#[path = "./helpers/lib.rs"]
+pub mod helpers;
 #[cfg(test)]
 #[path = "./testing/lib.rs"]
 pub mod tests_folder;
