+++ contracts/sg-eth-airdrop/src/msg.rs
@@ -14,11 +14,6 @@ pub struct InstantiateMsg {
     pub per_address_limit: u32,
 }
 
-// impl InstantiateMsg {
-//     fn into_wasm_msg() {}
-
-// }
-
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, Eq, JsonSchema)]
 pub struct AirdropClaimResponse {
     result: bool,
