+++ contracts/sg-eth-airdrop/src/query.rs
@@ -1,6 +1,7 @@
 use crate::{state::CONFIG, ContractError};
-use cosmwasm_std::DepsMut;
+use cosmwasm_std::{Deps, DepsMut, StdResult};
 use vending_minter::helpers::MinterContract;
+use whitelist_generic::helpers::WhitelistGenericContract;
 
 pub fn query_collection_whitelist(deps: &DepsMut) -> Result<String, ContractError> {
     let config = CONFIG.load(deps.storage)?;
@@ -12,5 +13,16 @@ pub fn query_collection_whitelist(deps: &DepsMut) -> Result<String, ContractErro
             None => Err(ContractError::CollectionWhitelistMinterNotSet {}),
         },
         Err(_) => Err(ContractError::CollectionWhitelistMinterNotSet {}),
+    }
+}
+
+pub fn query_airdrop_is_eligible(deps: Deps, eth_address: String) -> StdResult<bool> {
+    let config = CONFIG.load(deps.storage)?;
+    match config.whitelist_address {
+        Some(address) => WhitelistGenericContract(deps.api.addr_validate(&address)?)
+            .includes(&deps.querier, eth_address),
+        None => Err(cosmwasm_std::StdError::NotFound {
+            kind: "Whitelist Contract".to_string(),
+        }),
     }
 }
