+++ contracts/sg-eth-airdrop/src/testing/tests/test_claim.rs
@@ -6,6 +6,7 @@ use cosmwasm_std::{Addr, Attribute, Coin, Uint128};
 use ethers_core::rand::thread_rng;
 use ethers_signers::{LocalWallet, Signer};
 
+use crate::helpers::INSTANTIATION_FEE;
 use crate::tests_folder::constants::{
     MOCK_AIRDROP_ADDR_STR, MOCK_MINTER_ADDR_STR, NATIVE_DENOM, OWNER, STARGAZE_WALLET_01,
     STARGAZE_WALLET_02,
@@ -24,7 +25,7 @@ fn test_instantiate() {
     let (_, _, _, eth_addr_str) = get_wallet_and_sig(claim_plaintext.clone());
     let params = InstantiateParams {
         addresses: vec![eth_addr_str],
-        funds_amount: WHITELIST_AMOUNT,
+        funds_amount: WHITELIST_AMOUNT + INSTANTIATION_FEE,
         expected_airdrop_contract_id: 1,
         minter_address,
         admin_account: Addr::unchecked(OWNER),
@@ -45,7 +46,7 @@ fn test_valid_eth_sig_claim() {
 
     let params = InstantiateParams {
         addresses: vec![eth_addr_str.clone()],
-        funds_amount: WHITELIST_AMOUNT,
+        funds_amount: WHITELIST_AMOUNT + INSTANTIATION_FEE,
         expected_airdrop_contract_id: 4,
         minter_address: minter_addr.clone(),
         admin_account: Addr::unchecked(OWNER),
@@ -106,7 +107,7 @@ fn test_invalid_eth_sig_claim() {
 
     let params = InstantiateParams {
         addresses: vec![eth_addr_str.clone()],
-        funds_amount: WHITELIST_AMOUNT,
+        funds_amount: WHITELIST_AMOUNT + INSTANTIATION_FEE,
         expected_airdrop_contract_id: 4,
         minter_address: minter_addr.clone(),
         admin_account: Addr::unchecked(OWNER),
@@ -165,7 +166,7 @@ fn test_can_not_claim_twice() {
 
     let params = InstantiateParams {
         addresses: vec![eth_addr_str.clone()],
-        funds_amount: WHITELIST_AMOUNT,
+        funds_amount: WHITELIST_AMOUNT + INSTANTIATION_FEE,
         expected_airdrop_contract_id: 4,
         minter_address: minter_addr.clone(),
         admin_account: Addr::unchecked(OWNER),
@@ -233,7 +234,7 @@ fn test_claim_one_valid_airdrop() {
 
     let params = InstantiateParams {
         addresses: vec![eth_addr_str.clone()],
-        funds_amount: WHITELIST_AMOUNT,
+        funds_amount: WHITELIST_AMOUNT + INSTANTIATION_FEE,
         expected_airdrop_contract_id: 4,
         minter_address: minter_addr,
         admin_account: Addr::unchecked(OWNER),
@@ -282,7 +283,7 @@ fn test_claim_twice_receive_funds_once() {
 
     let params = InstantiateParams {
         addresses: vec![eth_addr_str.clone()],
-        funds_amount: WHITELIST_AMOUNT,
+        funds_amount: WHITELIST_AMOUNT + INSTANTIATION_FEE,
         expected_airdrop_contract_id: 4,
         minter_address: minter_addr,
         admin_account: Addr::unchecked(OWNER),
@@ -345,7 +346,7 @@ fn test_ineligible_does_not_receive_funds() {
 
     let params = InstantiateParams {
         addresses: vec![eth_addr_str],
-        funds_amount: WHITELIST_AMOUNT,
+        funds_amount: WHITELIST_AMOUNT + INSTANTIATION_FEE,
         expected_airdrop_contract_id: 4,
         minter_address: minter_addr,
         admin_account: Addr::unchecked(OWNER),
@@ -400,7 +401,7 @@ fn test_one_eth_claim_two_stargaze_addresses_invalid() {
 
     let params = InstantiateParams {
         addresses: vec![eth_addr_str_1.clone()],
-        funds_amount: WHITELIST_AMOUNT,
+        funds_amount: WHITELIST_AMOUNT + INSTANTIATION_FEE,
         expected_airdrop_contract_id: 4,
         minter_address: minter_addr.clone(),
         admin_account: Addr::unchecked(OWNER),
@@ -480,7 +481,7 @@ fn test_two_claims_allowed_success() {
 
     let params = InstantiateParams {
         addresses: vec![eth_addr_str.clone()],
-        funds_amount: WHITELIST_AMOUNT * 2,
+        funds_amount: WHITELIST_AMOUNT * 2 + INSTANTIATION_FEE,
         expected_airdrop_contract_id: 4,
         minter_address: minter_addr,
         admin_account: Addr::unchecked(OWNER),
